requirements.txt
PyPDF2==3.0.1
streamlit==1.29.0
pdfminer.six==20221105
google-generativeai==0.4.0
python-dotenv==1.0.0
protobuf==4.25.1